Fix the tainted session bug
Tainted session bug:
On a page refresh, the session cookie often (but not always) failed
to be overwritten by the /case/ and /reflection/ views. This caused a
token discrepancy on the next @session_protected request. The 't'
form field in the $.ajax request data would be the most up-to-date
version, while the session cookie in the header would still contain
an older version, causing verification to fail when it shouldn't.

Solution:
Let the @session_enable view decorator only set the token if there
isn't one in the session yet. Check whether a token was included on
the client side upon successful response.

In addition:
Only @session_enable /reflection/, not /case/. This in order to
prevent the theoretical possibility of a session race condition.