Fix the empty casus bug
When no casus archive is stored locally yet, the plate page fsm would
insist that the user should connect to the internet, even though the
data were just downloaded by the plate archive page fsm. The cause
was the fact that the "handled" event fires synchronously before the
data are stored asynchronously. The solution in this commit is to
emit another event when the data are actually stored, and make the
plate page fsm listen for it.